feat: modification des process d'authentification api pour prendre en compte des paramètres moins évidents que ceux liés au profile utilisateur

+++ class/Src/Api/Image.php
@@ -9,18 +9,22 @@ class Image
   {
     if ($isApi) {
 
-      \Src\Api\Auth::protect();
+      if (\Src\Api\Auth::protect()) {
 
-      $path = ROOT . "/upload/" . $table . "/" . $folderName . "/" . $subfolder . "/" . $fileName;
-      if (file_exists($path)) {
-        $imageData = base64_encode(file_get_contents($path));
-        echo json_encode($imageData);
-      } else {
-        http_response_code(404);
-        echo json_encode(['error' => 'Image not found']);
+        $path = ROOT . "/upload/" . $table . "/" . $folderName . "/" . $subfolder . "/" . $fileName;
+        if (file_exists($path)) {
+          $imageData = base64_encode(file_get_contents($path));
+          echo json_encode($imageData);
+        } else {
+
+          http_response_code(404);
+          echo json_encode(['error' => 'Image not found']);
+
+        }
+
+        return;
+
       }
-
-      return;
     }
 
     echo json_encode("403");
